Use quint64 for connection ID instead of int

+++ src/network/QxtRPCPeer.h
@@ -81,7 +81,7 @@ public:
      *
      * Servers must provide a client ID, provided by the \a clientConnected() signal; clients and peers must not.
      */
-    void disconnectPeer(int id = -1);
+    void disconnectPeer(quint64 id = -1);
     /*!
      * Disconnects from all clients, or from the server or peer.
      */
@@ -94,7 +94,7 @@ public:
     /*!
      * Returns a list of client IDs for all connected clients.
      */
-    QList<int> clients() const;
+    QList<quint64> clients() const;
 
     /*!
      * Attaches the given signal. 
@@ -138,7 +138,7 @@ public slots:
      * The receivers are not obligated to act upon the signal. If no client is connected with a provided ID, the ID
      * is ignored with a warning.
      */
-    void callClientList(QList<int> ids, QString fn, QVariant p1 = QVariant(), QVariant p2 = QVariant(), QVariant p3 = QVariant(), QVariant p4 = QVariant(),
+    void callClientList(QList<quint64> ids, QString fn, QVariant p1 = QVariant(), QVariant p2 = QVariant(), QVariant p3 = QVariant(), QVariant p4 = QVariant(),
               QVariant p5 = QVariant(), QVariant p6 = QVariant(), QVariant p7 = QVariant(), QVariant p8 = QVariant());
     /*!
      * Sends the signal fn with the given parameter list to the specified client.
@@ -147,7 +147,7 @@ public slots:
      * 
      * The receiver is not obligated to act upon the signal. If no client with the given ID is connected, the call will be ignored with a warning.
      */
-    void callClient(int id, QString fn, QVariant p1 = QVariant(), QVariant p2 = QVariant(), QVariant p3 = QVariant(), QVariant p4 = QVariant(),
+    void callClient(quint64 id, QString fn, QVariant p1 = QVariant(), QVariant p2 = QVariant(), QVariant p3 = QVariant(), QVariant p4 = QVariant(),
               QVariant p5 = QVariant(), QVariant p6 = QVariant(), QVariant p7 = QVariant(), QVariant p8 = QVariant());
     /*!
      * Sends the signal fn with the given parameter list to all connected clients except for the client specified.
@@ -157,7 +157,7 @@ public slots:
      * The receiver is not obligated to act upon the signal. This function is useful for rebroadcasting a signal from one client
      * to all other connected clients.
      */
-    void callClientsExcept(int id, QString fn, QVariant p1 = QVariant(), QVariant p2 = QVariant(), QVariant p3 = QVariant(), QVariant p4 = QVariant(),
+    void callClientsExcept(quint64 id, QString fn, QVariant p1 = QVariant(), QVariant p2 = QVariant(), QVariant p3 = QVariant(), QVariant p4 = QVariant(),
               QVariant p5 = QVariant(), QVariant p6 = QVariant(), QVariant p7 = QVariant(), QVariant p8 = QVariant());
     /*!
      * Detaches all signals and slots for the object that emitted the signal connected to detachSender().
@@ -174,7 +174,7 @@ signals:
      *
      * The given ID is used for disconnectPeer(), callClient(), and related functions.
      */
-    void clientConnected(int id);
+    void clientConnected(quint64 id);
     /*!
      * This signal is emitted when a peer or server is disconnected.
      */
@@ -182,7 +182,7 @@ signals:
     /*!
      * This signal is emitted when a client disconnects. The given ID is no longer valid.
      */
-    void clientDisconnected(int id);
+    void clientDisconnected(quint64 id);
     /*!
      * This signal is emitted whenever an error occurs on a socket.
      *
